feat(checker): Adds semantic check for nil usage
Introduces a new checker, NilChecker, to enforce the semantic rules for using the nil keyword as defined by the Tiger language specification.

According to the spec, nil can only be used in contexts where its type can be determined. This is limited to comparisons with, or assignments to, variables of a record type.